Extract env subcommand to cmd_env.go
Move env flags, command definition, and run function to separate file.
